fixed some shadows

+++ src/fs_standard_functions.h
@@ -24,8 +24,8 @@ f32 reverse_megabytes (u64 size) { return reverse_kilobytes(size) / 1024.0f; }
 f32 reverse_gigabytes (u64 size) { return reverse_megabytes(size) / 1024.0f; }
 f32 reverse_terabytes (u64 size) { return reverse_gigabytes(size) / 1024.0f; }
 
-template<typename _, u32 Count>
-constexpr u32 array_count(const _ (&array)[Count])
+template<typename _, s64 Count>
+constexpr s64 array_count(const _ (&array)[Count])
 {
     return Count;
 }
